test: use metallb in upgrade test instead of port-forward (#1259)
**Description**

This is a follow up on #1248 and #1255. This switches to use metallb
instead of port-forwarding multiple times to avoid the complexity/race
as well as to make the situation closer to the real world scenario where
type: LoadBalancer Service is utilized to access the gateway rather than
directly accessing pod ips via port-forward.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>