controller: run extproc as a sidecar container if k8s > v1.32 (#1263)
**Description**

This is a follow up on #1248. We need to ensure that extproc outlives
Envoy process and that can be easily achieved by configuring the extproc
container as a sidecar container in the k8s API level. Since that
feature is only available in k8s newer than v1.32, this adds a simple
logic to check the version to switch the container insertion to ensure
the old deployment works as intended.

---------

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>