feat: adds complete MCP proxy implementation (#1260)
**Description**

Adds support for Model Context Protocol (MCP). This implementation fully
implements version `2025-06-18 (latest)` of the MCP spec for the
Streamable HTTP transport and the OAuth Authorization.

For details about the implementation, architecture, and design choices,
please refer to the design proposal doc in the commit.

This has been a collective work by @mathetake @zhaohuabing @zirain and
@nacx.

**Related Issues/PRs (if applicable)**

Closes #589

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>
Co-authored-by: Takeshi Yoneda <t.y.mathetake@gmail.com>
Co-authored-by: Huabing Zhao <zhaohuabing@gmail.com>
Co-authored-by: zirain <zirain2009@gmail.com>