feat: add spanRequestHeaderAttributes and ENV mapping for CLI (#1269)
**Description**

This uses the same header-to-attribute mapping approach for both
OpenTelemetry spans and metrics, enabling session tracking and custom
attribute propagation without requiring code instrumentation.

Specifically, this deprecates `metricsRequestHeaderLabels` which was
artificially Prometheus specific in favor of:

**Kubernetes/extproc**
- `spanRequestHeaderAttributes`:
"x-session-id:session.id,x-user-id:user.id"
- `metricsRequestHeaderAttributes`:
"x-team-id:team.id,x-user-id:user.id"

Note: Before we told people to lower_snake_case label values, but you
don't need to do that because the prometheus exporter already does that.
We shouldn't make mapping decisions like this in aigw as it interferes
with non-prometheus metrics systems. This is particularly highlighted in
"session.id" which is handled where if you made it "session_id" it
wouldn't because the latter isn't an otel convention.

**Related Issues/PRs**

Fixes #1221

---------

Signed-off-by: Adrian Cole <adrian@tetrate.io>