fix: makes gRPC's MaxRecvMsgSize for the extension server (#1291)
**Description**
Makes go's gRPC `MaxRecvMsgSize` configurable for the extension server,
whose default value is 4MB
https://pkg.go.dev/google.golang.org/grpc#MaxRecvMsgSize

When the gRPC message payload from the xDS server exceeds 4MB, it is
rejected by the extension server and the below errors is observed on
envoy-gateway:
```
2025-10-07T10:26:44.625Z        ERROR   xds     runner/runner.go:245    failed to translate xds ir      {"runner": "xds", "error": "rpc error: code = ResourceExhausted desc = grpc: received message larger than max (6683527 vs. 4194304)"}
2025-10-07T10:26:44.625Z        ERROR   watchable       message/watchutil.go:86 observed an error       {"runner": "xds", "error": "rpc error: code = ResourceExhausted desc = grpc: received message larger than max (6683527 vs. 4194304)"}
```
This PR makes it configurable via the command-line flag `maxRecvMsgSize`
and keeps its default at 4MB.


**Related Issues/PRs (if applicable)**
Fixes https://github.com/envoyproxy/ai-gateway/issues/1293
Related PR to make gRPC's message sizes configurable, though that was
for the external processor gRPC server:
https://github.com/envoyproxy/ai-gateway/pull/1213


**Special notes for reviewers (if applicable)**

Note that this applies to the xDS resource messages sent from the xDS
server to the extension server. The payload of these messages increases
with the size/count of the config resources.

---------

Signed-off-by: Owayss Kabtoul <owayssk@gmail.com>