refactor: remove /crds folder from ai-gateway-helm chart (#1431)
**Description**

As discussed in this slack thread:
https://envoyproxy.slack.com/archives/C07Q4N24VAA/p1761235915409479, we
want to separate the crd chart and the resources chart. (option 1 of the
proposed approaches in
https://github.com/envoyproxy/ai-gateway/issues/1186)

**Related Issues/PRs (if applicable)**

https://github.com/envoyproxy/ai-gateway/issues/1186

**Special notes for reviewers (if applicable)**

Added an extra note on upgrade when done from a previous install of
ai-gateway-helm helm chart alone. it uses the `--take-ownership` flag
while upgrading: https://github.com/helm/helm/pull/13439,
https://helm.sh/docs/helm/helm_upgrade/#options

Open Question:
- Should we rename `ai-gateway-helm` to `ai-gateway-resources-helm`? It
would be more consistent but upgrade is slightly trickier

cc: @mathetake @johnugeorge @nagar-ajay

Signed-off-by: sailesh duddupudi <saileshradar@gmail.com>