test: make rate limit e2e less flaky (#1452)
**Description**
Since #1410, the deployment created by EG for the rate limit server is
in crashing loop until redis server is up and running in the rate limit
e2e test setup, which caused a bit of flake in the test. This fixes it
by restarting the deployment just in case.

This also removes the only remaining use of t.Context in test cleanup
function in e2e tests.

Signed-off-by: Takeshi Yoneda <t.y.mathetake@gmail.com>