feat: support response_format for gemini models other than 2.5 via translation from json schema to genai.Schema (#1400)
**Description**
Current implementation of response_format for gemini models uses
`responseJsonSchema`. However, in the official documentation, it's only
supported for gemini2.5:
https://ai.google.dev/gemini-api/docs/structured-output#json-schema.
Thus, for other versions of gemini models, we need to use
``responseSchema`. Because the supported fields of `responseSchema` are
different than standard json schema, we also need to add translations
from json schema to `responseSchema`

**Related Issues/PRs (if applicable)**
https://github.com/envoyproxy/ai-gateway/issues/1403.


In this pr, we only do translations for response_format. `Parameters` in
`genai.FunctionDeclaration` is also a `genai.Schema`, but it's typically
not complex and we would add translations for that late

---------

Signed-off-by: yxia216 <yxia216@bloomberg.net>
Signed-off-by: hustxiayang <yangxiast@gmail.com>
Signed-off-by: Dan Sun <dsun20@bloomberg.net>
Co-authored-by: Xiaolin Lin <xlin158@bloomberg.net>
Co-authored-by: Dan Sun <dsun20@bloomberg.net>