cli: align on apigroup for backendtlspolicy (#1500)
**Description**

Fixes a small issue that happened in standalone mode and the discovery
of BackendTLS policies for the MCPRoute security configuration.
In the controller code we're using `gwapiv1` in all reconcilers and the
controller-runtime Client. That client is typed and works with the
specific version. The fakeClient used in standalone mode, populated the
BackendTLSPolicy as `gwapiv1alpha3`, causing the controller code to not
find it.

We could change the controller code and make sure all operations that
use the `client.Client` from teh controller runtime account for all
versions, but that would be a lot of repeated code, and the controller
consistently uses `gwapiv1` everywhere, so I think it is OK to jsut fix
the standalone mode to initialize its fakeClient with the right version
(note that we can still "read any version from the YAML file", as we
generically unmarshal to `Unstructured`; it's just a matter of how we
populate the data in the fakeClient)..


**Related Issues/PRs (if applicable)**

N/A

**Special notes for reviewers (if applicable)**

N/A

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>