mcp: proxy stdio servers in standalone mode (#1479)
**Description**

Opening as a draft to get early feedback.

When running in standalone mode, it allows running and proxying `stdio`
MCP servers. This PR only supports defining the servers in the
`servers.json` format, as for now, we don't want to introduce new nouns
for a feature that does not apply to Kubernetes.

With this change, the CLI can be started with a file like the following:

```
aigw run --mcp-json '{
  "mcpServers": {
    "everything": {
      "command": "npx",
      "args": ["-y","@modelcontextprotocol/server-everything"]
    },
    "filesystem": {
      "command": "npx",
      "args": ["-y","@modelcontextprotocol/server-filesystem"]
    }
  }
}'
```

**Related Issues/PRs (if applicable)**

Depends on https://github.com/envoyproxy/gateway/pull/7427 for a clean
implementation.

**Special notes for reviewers (if applicable)**

N/A

---------

Signed-off-by: Ignasi Barrera <ignasi@tetrate.io>
Signed-off-by: Ignasi Barrera <nacx@apache.org>